refactor: modify :domain:domain-rds to :domain module
