+++ domain/src/test/java/site/ymango/domain/DomainApplicationTests.java
@@ -1,10 +1,10 @@
-package site.ymango.domainrds;
+package site.ymango.domain;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class DomainRdsApplicationTests {
+class DomainApplicationTests {
 
   @Test
   void contextLoads() {
