+++ settings.gradle.kts
@@ -1,4 +1,4 @@
 rootProject.name = "flirting"
 
 include(":app:app-service-api")
-include(":domain:domain-rds")
+include(":domain")
